requirements.txt
aiohttp==3.8.5
pydantic==2.7.1
python-dotenv==1.0.1
telethon==1.35.0
